Revert "refactor(mantine): fold the chrome and table files into chrome.tsx and tables.tsx"
This reverts commit 384816e.

The merge cut file count without improving anything: 11 files became 2, the
line count did not move, and DesktopTable's 912 lines went from their own file
into a 1,346-line tables.tsx that is harder to navigate. Mantine's split was
the better layout; the shared standard is the one worth revisiting.
